Updated torque vectoring main file for first test running on car

+++ boards/DCM/Src/App/torquevectoring/App_PowerLimiting.c
@@ -15,10 +15,11 @@ float App_PowerLimiting_ComputeMaxPower(struct PowerLimiting_Inputs *inputs)
     // ============== Calculate max powers =================
     // 1. Motor Temps
     float P_max_motor_temps = POWER_LIMIT_CAR_kW;
-    if (max_motor_temp - MOTOR_TEMP_CUTOFF_c >= 30.0)
+    if ((float)max_motor_temp - (float)MOTOR_TEMP_CUTOFF_c >= (float)30.0)
     {
         P_max_motor_temps = 0.0;
-    } else if (max_motor_temp > MOTOR_TEMP_CUTOFF_c)
+    }
+    else if (max_motor_temp > MOTOR_TEMP_CUTOFF_c)
     {
         P_max_motor_temps =
             POWER_LIMIT_CAR_kW - (max_motor_temp - MOTOR_TEMP_CUTOFF_c) * MOTOR_TEMP_POWER_DECREMENTING_RATIO;
